Check if diseaseNames collection exists before trying to drop

+++ create_disease_label_collection.py
@@ -4,6 +4,9 @@ import config
 
 if __name__ == '__main__':
     db = MongoClient('%s/girder' % config.mongo_url)
-    db.drop_collection('diseaseNames')
+
+    if 'diseaseNames' in db.collection_names():
+      db.drop_collection('diseaseNames')
+
     for label in disease_label_table.get_labels():
         db['diseaseNames'].insert({'_id': label})
